+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 # [8.14.0] - 2024-11-??
 - Close HTTP responses to prevent resource leaks
 - Added `RedactResponseException` and deprecated `VonageBadRequestException`
+- Added `maxBitrate` to `com.vonage.client.video.Archive`
 
 # [8.13.0] - 2024-10-28
 - Added support for Verify custom templates